GAL-178-Create-GitHub-repo-for-each-module
This commit is to temporarily disable the check-file-format and
check-markdown-format to avoid forcing external libraries and packages to
follow our format. This will also reduce the amount of files the
script needs to screen for allowing quicker and more relavant checks

A ticket out of the back of this is needed to investigate adding
exclusions to directors and files that will not need formatting-format actions on the pipeline.

The reason behind this is because exclusions need to be added in
in order to allow for the above two stages to only screen the files
we create and avoid forcing external libraries and packages to
follow our format. This will also reduce the amount of files the
script needs to screen for allowing quicker and more relavant checks

A ticket out of the back of this is needed to investigate adding
exclusions to directors and files that will not need formatting

GAL-178-Create-GitHub-repo-for-each-module